requirements.txt
asgiref==3.2.3
Django==3.0.3
django-cleanup==4.0.0
django-crispy-forms==1.9.0
Pillow==7.1.1
pytz==2019.3
pyuca==1.2
sqlparse==0.3.0